+++ src/be/tarsos/lsh/families/CityBlockHashFamily.java
@@ -45,8 +45,8 @@ public class CityBlockHashFamily implements HashFamily {
 	}
 
 	@Override
-	public Integer combine(int[] hashes) {
-		return Arrays.hashCode(hashes);
+	public Long combine(int[] hashes) {
+		return (long) Arrays.hashCode(hashes);
 	}
 
 	@Override
